accounts/requirements.txt
Django==5.0.7
pillow
crispy-bootstrap5
django-crispy-forms
djangorestframework
requests


